corrected my cookie setting

+++ server/controllers/user.controller.js
@@ -5,9 +5,9 @@ import jwt from "jsonwebtoken";
 
 export const updateUser = async (req, res, next) => {
   console.log(req.userInfo);
-  // if (req.userInfo.userID !== req.params.userId) {
-  //   return next(errorHandler(403, "You are not allowed to update this user"));
-  // }
+  if (req.userInfo.userID !== req.params.userId) {
+    return next(errorHandler(403, "You are not allowed to update this user"));
+  }
   if (req.body.password) {
     if (req.body.password.length < 6) {
       return next(
@@ -66,9 +66,9 @@ export const updateUser = async (req, res, next) => {
       { new: true }
     );
     const { password: pass, ...rest } = updatedUser._doc;
-    const token = jwt.sign({ userID: updateUser._id }, process.env.SECRETE);
-    res.status(201).cookie("access_token", token).json({ token, rest });
-    // res.status(201).json(rest);
+    // const token = jwt.sign({ userID: updateUser._id }, process.env.SECRETE);
+    // res.status(201).cookie("access_token", token).json({ token, rest });
+    res.status(201).json(rest);
   } catch (error) {
     next(error);
   }
@@ -77,11 +77,11 @@ export const updateUser = async (req, res, next) => {
 export const deleteUser = async (req, res, next) => {
   // console.log(req.userInfo);
 
-  // if (req.userInfo.userID !== req.params.userId) {
-  //   return next(
-  //     errorHandler(403, "You are not allowed to delete this account")
-  //   );
-  // }
+  if (req.userInfo.userID !== req.params.userId) {
+    return next(
+      errorHandler(403, "You are not allowed to delete this account")
+    );
+  }
 
   try {
     await User.findByIdAndDelete(req.params.userId);
